+++ crates/holochain/src/core/queue_consumer/app_validation_consumer.rs
@@ -3,6 +3,7 @@
 use super::*;
 use crate::core::workflow::app_validation_workflow::app_validation_workflow;
 use crate::core::workflow::app_validation_workflow::AppValidationWorkspace;
+use crate::core::workflow::sys_validation_workflow::get_representative_agent;
 
 /// Spawn the QueueConsumer for AppValidation workflow
 #[cfg_attr(
@@ -32,14 +33,22 @@ pub fn spawn_app_validation_consumer(
         conductor.task_manager(),
         (tx.clone(), rx),
         move || {
-            app_validation_workflow(
-                dna_hash.clone(),
-                workspace.clone(),
-                trigger_integration.clone(),
-                trigger_publish.clone(),
-                conductor.clone(),
-                network.clone(),
-            )
+            if let Some(representative_agent) =
+                get_representative_agent(&conductor, &network.dna_hash())
+            {
+                Either::Left(app_validation_workflow(
+                    dna_hash.clone(),
+                    workspace.clone(),
+                    trigger_integration.clone(),
+                    trigger_publish.clone(),
+                    conductor.clone(),
+                    network.clone(),
+                    representative_agent,
+                ))
+            } else {
+                tracing::warn!("No representative agent found for DNA, skipping app validation.");
+                Either::Right(async move { Ok(WorkComplete::Complete) })
+            }
         },
     );
     tx